added currency on the main screen

+++ app/src/main/java/com/example/tinkoffproject/model/data/dto/Currency.kt
@@ -4,5 +4,5 @@ data class Currency(
     val shortName: String,
     val longName: String,
     val isUp: Boolean,
-    val rate: Float
+    val rate: Double
 )
